requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.9.0
billiard==3.6.3.0
brotlipy==0.7.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-bootstrap4==1.1.1
django-celery-beat==2.0.0
django-compat==1.0.15
django-dia==0.4.1
django-extensions==3.0.2
django-filter==2.3.0
django-heroku==0.3.1
django-redis-cache==2.1.1
django-stdimage==5.0.2
django-timezone-field==4.0
djangorestframework==3.11.0
dnspython==1.16.0
geojson==2.5.0
gevent==20.5.0
graphviz==0.14
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
joblib==0.15.1
json-table-schema==0.2.1
kml2geojson==4.0.2
kombu==4.6.8
lazy-object-proxy==1.5.0
lxml==4.5.0
Markdown==3.2.2
mccabe==0.6.1
messytables==0.15.2
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
monotonic==1.5
mysql==0.0.2
mysql-client==0.0.1
mysqlclient==1.3.14
numpy==1.18.1
olefile==0.46
pandas==1.0.3
Pillow==6.2.1
pisa==3.0.33
progressbar2==3.50.1
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.5
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-magic==0.4.15
python-utils==2.4.0
pytz==2020.1
redis==3.5.2
reportlab==3.5.42
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
tinycss2==1.0.2
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.25.9
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
wget==3.2
whitenoise==5.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
zipp==3.1.0
